mock-gateway/gateway-rower-controller/requirements.txt
importlib_metadata<5
numpy
pandas
# --extra-index-url http://${HOST_DOCKER_INTERNAL}:${LOCAL_PYPI_PORT}/simple/
--trusted-host ${HOST_DOCKER_INTERNAL}
# smartgymgatewayutils
requests
redis
paho-mqtt
celery